Consolidate requirements files and reorganize software files
Requirements consolidation:
- Consolidate 6 requirements files to 3 files
- Create unified requirements-pi.txt (works for both 32-bit and 64-bit Pi)
- Remove redundant requirements.txt, requirements_32bit.txt, requirements_64bit.txt
- Remove requirements-webapp-only-32bit.txt (replaced by requirements-pi.txt)
- Keep requirements-simulation.txt and requirements-api.txt
- Update all script and documentation references

File organization:
- Move flow_control_modes.py to controllers/ (flow-control specific)
- Move omw.desktop to rio-webapp/ (desktop shortcut for webapp)
- Update config.py import path for flow_control_modes
- Update deployment scripts to copy from new locations
- Document flow_control_modes.py in controllers/README.md
- Document omw.desktop in rio-webapp/README.md

Update okh-manifest.yml with new requirements file path
