+++ software/requirements-pi.txt
@@ -1,10 +1,11 @@
-# Requirements for Rio Webapp Only (32-bit Raspberry Pi)
-# This file is for Raspberry Pi deployment - install to SYSTEM Python (no virtual environment)
-# Hardware packages (spidev, RPi.GPIO, picamera) should be installed via apt (system-wide)
+# Requirements for Rio Controller - Raspberry Pi Deployment
+# This file works for both 32-bit and 64-bit Raspberry Pi OS
+# Install to SYSTEM Python (no virtual environment)
+# Hardware packages (spidev, RPi.GPIO, picamera/picamera2) should be installed via apt (system-wide)
 
-# Web framework (upgrades from existing versions)
+# Web framework (compatible with both 32-bit and 64-bit)
 Flask>=2.0.0,<4.0.0
-Flask-SocketIO>=5.4.0,<6.0.0  # Upgraded from 4.3.2 - tested and working
+Flask-SocketIO>=5.4.0,<6.0.0
 Werkzeug>=2.0.0,<4.0.0
 Jinja2>=3.0.0
 MarkupSafe>=2.0.0
@@ -13,9 +14,8 @@ gevent>=23.0.0,<25.0.0
 gevent-websocket>=0.10.1
 
 # WebSocket support (compatible with Socket.IO client v2/v3/v4)
-# Upgraded versions compatible with Flask-SocketIO 5.4+
 python-socketio>=5.14.0
-python-engineio>=4.9.0  # Upgraded from 3.13.2 - Version 4.x required for python-socketio 5.x
+python-engineio>=4.9.0  # Version 4.x required for python-socketio 5.x
 
 # Eventlet for async support (required for Flask-SocketIO)
 eventlet>=0.33.0,<1.0.0
@@ -25,6 +25,7 @@ eventlet>=0.33.0,<1.0.0
 # Do NOT install opencv-python-headless from pip - it builds from source (very slow, hours).
 # If python3-opencv is not available, uncomment the line below (but expect long build time):
 # opencv-python-headless<4.9  # Only if python3-opencv from apt is not available
+# Note: numpy<2.0.0 required for 32-bit compatibility (numpy 2.0+ doesn't work on 32-bit)
 numpy<2.0.0  # Prefer apt: sudo apt-get install -y python3-numpy
 Pillow>=9.0.0  # Required for image processing in camera modules
 
@@ -33,15 +34,17 @@ PyYAML>=6.0
 
 # Note: The following system packages should be installed via apt BEFORE installing Python packages:
 # Hardware packages:
-#   sudo apt-get install -y python3-spidev python3-rpi.gpio python3-picamera python3-numpy python3-opencv
+#   For 32-bit: sudo apt-get install -y python3-spidev python3-rpi.gpio python3-picamera python3-numpy python3-opencv
+#   For 64-bit: sudo apt-get install -y python3-spidev python3-rpi.gpio python3-picamera2 python3-numpy python3-opencv
 # BLAS/LAPACK for droplet detection (provides libcblas.so.3):
 #   sudo apt-get install -y libatlas-base-dev libatlas3-base libblas3 liblapack3
 # IMPORTANT: Install python3-opencv from apt (fast, pre-built). Do NOT install opencv-python-headless from pip.
 #
 # Installation instructions for Raspberry Pi:
-# Step 1: Install system packages (see above)
+# Step 1: Install system packages (see above - choose picamera or picamera2 based on your OS)
 # Step 2: Install Python packages to system Python (NO virtual environment):
 #   python3 -m pip install --user --upgrade pip wheel
-#   python3 -m pip install --user -r requirements-webapp-only-32bit.txt
+#   python3 -m pip install --user -r requirements-pi.txt
 # Or with sudo for system-wide installation:
-#   sudo python3 -m pip install -r requirements-webapp-only-32bit.txt
+#   sudo python3 -m pip install -r requirements-pi.txt
+
